Added CellOption trait to create common interface for Cell behaviour on right click
currently we have two parallel vectors for retaining displayed wave
state within the sigwindow module.

There is the live_waves vector, which gets fed into the wavewindow
component (the component that visualizes the waves) and then there is
the waves_state CellList that generates the names of the signals that
are displayed.

I wanted to move this state to all be held by the CellList, but borrow
checker has crushed my dreams. I am no longer interested in making this
clean, we can have parallel state for now