fixing issues with puddle. commit includes;
* basic support for zx signals
* patching issues with signals of length % 8
* unit tests for puddle builder, pcursor and droplet interfaces
* integration tests with wavedb and in memory waves

currently the stack of abstractions for signals looks like (from top to
bottom)

1) Inmemwave (wave2 interface)
2) PCursor
3) Puddle/Droplet (storage interface)

currently an in memory wave simply pulls ALL puddles for a given signal
into memory. The only interface to access underlying data is the
all_data() method, that creates an iterator that streams through all of
those puddles.

Things to think about still.

1) Limited access to signal (think windowing, going from (a,b))
2) random access of to a signal