cleaned up the puddle builder logic surrounding extending data from Vec<vcd::Value>.
Previously we took data as coming in big endian, with bits reversed.
Caused a bit of nasty logic to cope with this. Turns out if we just
reverse the vector then we can treat everything as little endian and we
have a bitstream coming in where the first bit  == bit 0 of the signal,
second bit is bit 1, etc etc

This whole endevaour kind of illuminated how coupled this project is to
the vcd crate. It is the only way we ingest data, and there isn't a
clear interface that could be swapped out. I think it would be valuable
to develop some trait to better capture expected behaviour coming from a
VCD parser, so we aren't untangling gross logic later down the line